BAC Calculation

+++ AlcoholApp/src/AlcoholApp/ViewModels/AlcoholDTO.cs
@@ -12,7 +12,7 @@ namespace AlcoholApp.ViewModels
         public string Type { get; set; }
         public string Brand { get; set; }
         public string Style { get; set; }
-        public decimal ABV { get; set; }
+        public double ABV { get; set; }
         public List<decimal> Volumes { get; set; }
         public ICollection<GlassDTO> Glasses { get; set; }
